代理+微步+网站截图/汇总代码/requirements.txt
beautifulsoup4
Flask
loguru
opencv_python
Pillow
pyzbar
Requests
selenium
selenium_wire
urllib3
webdriver_manager
